+++ eth_defi/erc_4626/classification.py
@@ -1095,6 +1095,11 @@ def create_vault_instance(
 
         return EthStrategyVault(web3, spec, token_cache=token_cache, features=features)
 
+    elif ERC4626Feature.yuzu_money_like in features:
+        from eth_defi.erc_4626.vault_protocol.yuzu_money.vault import YuzuMoneyVault
+
+        return YuzuMoneyVault(web3, spec, token_cache=token_cache, features=features)
+
     else:
         # Generic ERC-4626 without fee data
         from eth_defi.erc_4626.vault import ERC4626Vault
@@ -1174,6 +1179,9 @@ HARDCODED_PROTOCOLS = {
     # ETH Strategy - ESPN (EthStrategyPerpetualNote) vault on Ethereum
     # https://etherscan.io/address/0xb250c9e0f7be4cff13f94374c993ac445a1385fe
     "0xb250c9e0f7be4cff13f94374c993ac445a1385fe": {ERC4626Feature.eth_strategy_like},
+    # Yuzu Money - yzPP (Yuzu Protection Pool) vault on Plasma
+    # https://plasmascan.to/address/0xebfc8c2fe73c431ef2a371aea9132110aab50dca
+    "0xebfc8c2fe73c431ef2a371aea9132110aab50dca": {ERC4626Feature.yuzu_money_like},
 }
 
 for a in HARDCODED_PROTOCOLS.keys():