+++ eth_defi/erc_4626/vault_protocol/yuzu_money/__init__.py
@@ -1,0 +1,1 @@
+"""Yuzu Money protocol integration."""